control|loader: fix module unloading order, could have led to segfault before
It is preferred to do a cleanup before we unload a library. That way we can
actually purge module data and free its occupied memory. If we try it after
unloading the respective library, we would most likely segfault, since when
unloading the library, the corresponding memory was cleaned with it, and
its destructors etc won't be available to us anymore.